requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi[standard]==0.123.4
psutil==7.1.3
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
supabase==2.24.0
uvicorn[standard]==0.34.0